requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
SQLAlchemy==2.0.23
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
reportlab==4.0.7
Pillow==10.1.0
cryptography==41.0.7
bcrypt==4.1.2
PyJWT==2.8.0
requests==2.31.0
python-dateutil==2.8.2
pymysql==1.1.0
psutil==5.9.6 
